requirements.txt
flet>=0.24
anthropic>=0.40
google-generativeai>=0.8
